+++ CHANGELOG.md
@@ -25,6 +25,7 @@ minor release, the component will be purged, so be prepared (see `Updating` sect
 - Sessions to RPC server running in IR's local consensus mode (#2532)
 - `neofs-cli object nodes` command to get SNs for an object (#2512)
 - Fetching container estimations via iterators to prevent NeoVM stack overflow (#2173)
+- `neofs-adm morph netmap-candidates` CLI command (#1889)
 
 ### Fixed
 - `neo-go` RPC connection loss handling (#1337)